refactor(go-sdk): move v3 handlers into dedicated `handlers/v3` package
