+++ projects/go-sdk/handlers/v3/deleteVisitorData.go
@@ -1,4 +1,4 @@
-package handlers
+package handlersv3
 
 import (
 	"encoding/json"
@@ -6,21 +6,22 @@ import (
 	"net/http"
 )
 
-type getRelatedVisitorsQueryParams struct {
+type deleteVisitorDataQueryParams struct {
 	ApiKey    string `json:"apiKey"`
 	Region    string `json:"region"`
 	VisitorId string `json:"visitorId"`
 }
 
-func GetRelatedVisitors(w http.ResponseWriter, r *http.Request) {
-	queryParams := getRelatedVisitorsQueryParams{
+func DeleteVisitorData(w http.ResponseWriter, r *http.Request) {
+	queryParams := deleteVisitorDataQueryParams{
 		ApiKey:    r.URL.Query().Get("apiKey"),
 		Region:    r.URL.Query().Get("region"),
 		VisitorId: r.URL.Query().Get("visitorId"),
 	}
 
 	client, auth := fingerprintv3.Init(queryParams.ApiKey, queryParams.Region)
-	response := fingerprintv3.ProcessResponse(client.FingerprintApi.GetRelatedVisitors(auth, queryParams.VisitorId))
+	httpRes, err := client.FingerprintApi.DeleteVisitorData(auth, queryParams.VisitorId)
+	response := fingerprintv3.ProcessResponse(nil, httpRes, err)
 
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(200)
